move action locator into separate file
