+++ src/components/BrowserEventList/EventDescription/EventDescription.tsx
@@ -1,4 +1,4 @@
-import { ActionLocator } from '@/main/runner/schema'
+import { ActionLocator } from '@/schemas/locator'
 import { BrowserEvent } from '@/schemas/recording'
 import { exhaustive } from '@/utils/typescript'
