+++ src/components/HighlightSelectorProvider.tsx
@@ -1,6 +1,6 @@
 import { createContext, ReactNode, useContext, useState } from 'react'
 
-import { ActionLocator } from '@/main/runner/schema'
+import { ActionLocator } from '@/schemas/locator'
 
 type SetHighlightedSelector = (selector: ActionLocator | null) => void
