+++ src/handlers/browserRemote/index.ts
@@ -1,6 +1,6 @@
 import { ipcMain } from 'electron'
 
-import { ActionLocator } from '@/main/runner/schema'
+import { ActionLocator } from '@/schemas/locator'
 
 import { BrowserRemoteHandlers } from './types'
