+++ src/main/runner/schema.ts
@@ -1,6 +1,8 @@
 import { EventType, type eventWithTime } from '@rrweb/types'
 import { z } from 'zod/v4'
 
+import { ActionLocatorSchema } from '@/schemas/locator'
+
 /**
  * Creates a fault-tolerant schema that returns `undefined` on failure. This is used
  * to guard against the user passing invalid options that would otherwise cause the entire
@@ -12,74 +14,6 @@ function safe<T>(schema: z.ZodType<T>) {
 
 // Generic options schema to allow any options object. Should be refined later.
 const GenericOptions = z.unknown()
-
-const CssLocatorSchema = z.object({
-  type: z.literal('css'),
-  selector: z.string(),
-})
-
-const GetByRoleLocatorSchema = z.object({
-  type: z.literal('role'),
-  role: z.string(),
-  options: z
-    .object({
-      name: z.string().optional(),
-      exact: z.boolean().optional(),
-    })
-    .optional(),
-})
-
-const GetByTestIdLocatorSchema = z.object({
-  type: z.literal('testid'),
-  testId: z.string(),
-})
-
-const TextLocatorOptions = z
-  .object({
-    exact: z.boolean().optional(),
-  })
-  .optional()
-
-const GetByAltTextLocatorSchema = z.object({
-  type: z.literal('alt'),
-  text: z.string(),
-  options: TextLocatorOptions,
-})
-
-const GetByLabelLocatorSchema = z.object({
-  type: z.literal('label'),
-  label: z.string(),
-  options: TextLocatorOptions,
-})
-
-const GetByPlaceholderLocatorSchema = z.object({
-  type: z.literal('placeholder'),
-  placeholder: z.string(),
-  options: TextLocatorOptions,
-})
-
-const GetByTitleLocatorSchema = z.object({
-  type: z.literal('title'),
-  title: z.string(),
-  options: TextLocatorOptions,
-})
-
-const GetByTextLocatorSchema = z.object({
-  type: z.literal('text'),
-  text: z.string(),
-  options: TextLocatorOptions,
-})
-
-export const ActionLocatorSchema = z.discriminatedUnion('type', [
-  CssLocatorSchema,
-  GetByRoleLocatorSchema,
-  GetByTestIdLocatorSchema,
-  GetByAltTextLocatorSchema,
-  GetByLabelLocatorSchema,
-  GetByPlaceholderLocatorSchema,
-  GetByTitleLocatorSchema,
-  GetByTextLocatorSchema,
-])
 
 const PageGotoActionSchema = z.object({
   method: z.literal('page.goto'),
@@ -347,8 +281,6 @@ export const BrowserActionEventSchema = z.discriminatedUnion('type', [
   ActionBeginEventSchema,
   ActionEndEventSchema,
 ])
-
-export type ActionLocator = z.infer<typeof ActionLocatorSchema>
 
 export type ActionBeginEvent = z.infer<typeof ActionBeginEventSchema>
 export type ActionEndEvent = z.infer<typeof ActionEndEventSchema>
