+++ src/schemas/locator.ts
@@ -1,0 +1,71 @@
+import { z } from 'zod/v4'
+
+const CssLocatorSchema = z.object({
+  type: z.literal('css'),
+  selector: z.string(),
+})
+
+const GetByRoleLocatorSchema = z.object({
+  type: z.literal('role'),
+  role: z.string(),
+  options: z
+    .object({
+      name: z.string().optional(),
+      exact: z.boolean().optional(),
+    })
+    .optional(),
+})
+
+const GetByTestIdLocatorSchema = z.object({
+  type: z.literal('testid'),
+  testId: z.string(),
+})
+
+const TextLocatorOptions = z
+  .object({
+    exact: z.boolean().optional(),
+  })
+  .optional()
+
+const GetByAltTextLocatorSchema = z.object({
+  type: z.literal('alt'),
+  text: z.string(),
+  options: TextLocatorOptions,
+})
+
+const GetByLabelLocatorSchema = z.object({
+  type: z.literal('label'),
+  label: z.string(),
+  options: TextLocatorOptions,
+})
+
+const GetByPlaceholderLocatorSchema = z.object({
+  type: z.literal('placeholder'),
+  placeholder: z.string(),
+  options: TextLocatorOptions,
+})
+
+const GetByTitleLocatorSchema = z.object({
+  type: z.literal('title'),
+  title: z.string(),
+  options: TextLocatorOptions,
+})
+
+const GetByTextLocatorSchema = z.object({
+  type: z.literal('text'),
+  text: z.string(),
+  options: TextLocatorOptions,
+})
+
+export const ActionLocatorSchema = z.discriminatedUnion('type', [
+  CssLocatorSchema,
+  GetByRoleLocatorSchema,
+  GetByTestIdLocatorSchema,
+  GetByAltTextLocatorSchema,
+  GetByLabelLocatorSchema,
+  GetByPlaceholderLocatorSchema,
+  GetByTitleLocatorSchema,
+  GetByTextLocatorSchema,
+])
+
+export type ActionLocator = z.infer<typeof ActionLocatorSchema>
