+++ src/views/BrowserTestEditor/ActionForms/forms/locators/GetByLabelForm.tsx
@@ -1,5 +1,5 @@
 import { FieldGroup } from '@/components/Form'
-import { ActionLocator } from '@/main/runner/schema'
+import { ActionLocator } from '@/schemas/locator'
 
 import { TextFieldWithExactToggle } from '../../components'
 import { toFieldErrors } from '../utils'
