+++ src/views/BrowserTestEditor/ActionForms/forms/locators/GetByTestIdForm.tsx
@@ -1,7 +1,7 @@
 import { TextField } from '@radix-ui/themes'
 
 import { FieldGroup } from '@/components/Form'
-import { ActionLocator } from '@/main/runner/schema'
+import { ActionLocator } from '@/schemas/locator'
 
 import { toFieldErrors } from '../utils'
