+++ src/views/Validator/Browser/BrowserActionLocator.tsx
@@ -1,5 +1,5 @@
 import { Locator } from '@/components/Browser/Locator'
-import { ActionLocator } from '@/main/runner/schema'
+import { ActionLocator } from '@/schemas/locator'
 
 import { useHighlightSelector } from '../../../components/HighlightSelectorProvider'
